+++ src/components/market/detail/image-slide.tsx
@@ -3,7 +3,7 @@ import { NextPage } from "next";
 import Image from "next/image";
 import { translateClasses } from "../../../lib/translate-class";
 import { cls } from "../../../utils/class";
-import LoadingSpinner from "../../loading-spinner";
+import LoadingSpinner from "../../ui/loading-spinner";
 import useSlide from "../../../hooks/useSlide";
 
 interface Images {
