+++ src/components/ui/header.tsx
@@ -2,7 +2,7 @@ import { NextPage } from "next";
 import { Icon } from "@iconify/react";
 import Link from "next/link";
 import Image from "next/image";
-import logo from "../../public/asset/image/logo.png";
+import logo from "../../../public/asset/image/logo.png";
 import { useRouter } from "next/router";
 import { signOut } from "next-auth/react";
