+++ src/pages/create/index.tsx
@@ -4,11 +4,11 @@ import React, { ChangeEvent, useEffect, useState } from "react";
 import { Icon } from "@iconify/react";
 import { useForm, FieldErrors } from "react-hook-form";
 import { useMutation } from "react-query";
-import Button from "../../components/button";
-import Header from "../../components/header";
+import Button from "../../components/ui/button";
+import Header from "../../components/ui/header";
 import UploadImages from "../../components/create/upload-images";
 import OptionTab from "../../components/create/option-tab";
-import LoadingSpinner from "../../components/loading-spinner";
+import LoadingSpinner from "../../components/ui/loading-spinner";
 import { cls } from "../../utils/class";
 import { createImageUrl } from "../../utils/image-url";
 import axios from "axios";
