+++ src/pages/mypage/profile.tsx
@@ -6,8 +6,8 @@ import React, { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useMutation } from "react-query";
 import { useRecoilRefresher_UNSTABLE, useRecoilValueLoadable } from "recoil";
-import Header from "../../components/header";
-import Navigation from "../../components/navigation";
+import Header from "../../components/ui/header";
+import Navigation from "../../components/ui/navigation";
 import UserManage from "../../components/user/manage";
 import useUpload from "../../hooks/useUpload";
 import { taglist } from "../../lib/tag-data";
